+++ packages/component-library/src/components/form/_internal/mt-select-base/_internal/mt-select-result.vue
@@ -4,6 +4,7 @@
     :class="resultClasses"
     @mouseenter="onMouseEnter"
     @click.stop="onClickResult"
+    :aria-disabled="disabled"
   >
     <span class="mt-select-result__result-item-preview">
       <slot name="preview" />