+++ packages/component-library/utils/svgStringifyTransformer.js
@@ -1,4 +1,4 @@
-// eslint-disable-next-line no-undef
+// eslint-disable-next-line no-undef, @typescript-eslint/no-require-imports
 const path = require("path");
 
 // eslint-disable-next-line no-undef