+++ packages/prettier-config/index.js
@@ -1,0 +1,7 @@
+/**
+ * @type {import("prettier").Config}
+ */
+export default {
+  trailingComma: "es5",
+  singleQuote: false,
+};