+++ packages/admin-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 5.7.5
+
+### Patch Changes
+
+- Updated dependencies [1b0b620]
+- Updated dependencies [6cc3876]
+  - @shopware-ag/meteor-component-library@4.4.1
+
 ## 5.7.4
 
 ### Patch Changes